Introduce DeprecatedClassHelper
